Remove legacy stage plan interpreter
